+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * Add `:allow_nil` and `:allow_blank` to `validation_options` plugin (@aradunovic)
 * Execute validation block in context of the Fend class (@aradunovic)
+* Add `contexts` plugin (@aradunovic)